+++ src/chalk.nim
@@ -45,6 +45,7 @@ when isMainModule:
   of "exec":               runCmdExec(getArgs())
   of "setup":              runCmdSetup()
   of "docgen":             runChalkDocGen() # in cmd_help
+  of "__.onbuild":         runCmdOnBuild() # in cmd_internal
   else:
     runChalkHelp(getCommandName()) # noreturn, will not show config.