+++ src/commands.nim
@@ -9,6 +9,7 @@
 ## --publish-defaults functionality for other commands.
 
 import commands/[cmd_insert, cmd_extract, cmd_delete, cmd_docker, cmd_exec,
-                 cmd_env, cmd_dump, cmd_load, cmd_version, cmd_setup, cmd_help]
+                 cmd_env, cmd_dump, cmd_load, cmd_version, cmd_setup, cmd_help,
+                 cmd_internal]
 export cmd_insert, cmd_extract, cmd_delete, cmd_docker, cmd_exec, cmd_env,
-       cmd_dump, cmd_load, cmd_version, cmd_setup, cmd_help
+       cmd_dump, cmd_load, cmd_version, cmd_setup, cmd_help, cmd_internal